fix issue with large number for big directs

+++ asm/include/asm.h
@@ -67,7 +67,7 @@
 
 typedef struct					s_argval
 {
-	int							value;
+	unsigned long long						value;
 	char						*label;
 }								t_argval;
